REFACTOR TestSourceConnection form to its own template
