+++ pwa/src/templates/sourcesDetailTemplate/TestSourceConnectionForm/TestSourceConnectionFormTemplate.module.css
@@ -1,0 +1,17 @@
+.grid {
+  display: grid;
+  grid-gap: var(--gateway-ui-size-lg);
+  grid-template-columns: 1fr 1fr;
+}
+
+.gridContainer > *:not(:first-child) {
+  margin-block-start: var(--gateway-ui-size-sm);
+}
+
+.testConnectionButton {
+  margin-block-end: var(--gateway-ui-size-lg);
+}
+
+.buttonIcon > span > svg {
+  margin-inline-end: var(--gateway-ui-size-xs);
+}
